+++ src/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -330,6 +330,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .eq("rollup.config.js"),
             ofFile("scalingo", "/icons/scalingo.svg", "Scalingo: scalingo.json")
                 .eq("scalingo.json"),
+            ofFile("security", "/icons/security.svg", "Security: security(.md,.txt,.adoc)")
+                .eq("security").mayEnd(TXT),
             ofFile("snapcraft", "/icons/snap.svg", "Snapcraft: snapcraft.yaml")
                 .eq("snapcraft.yaml"),
             ofFile("swaggerignore", "/icons/swaggerignore.svg", "Swagger: .swagger-codegen-ignore")