+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * support Prettier `.prettierignore` files.
 * support Scalingo `scalingo.json` files.
 * support Snapcraft `snapcraft.yaml` files.
+* support `SECURITY` files (with `.md`, `.adoc`, `.txt` or no extension).
 * support [Karate](https://github.com/intuit/karate) `karate-config.js` files.
 * support [Crowdin](https://crowdin.com) `.crowdin.yml` files.
 * rework 16x16 icon for `package-info.java` files.